@@ -21,14 +21,15 @@
 		"autoprefixer": "^10.4.16",
 		"eslint": "^8.56.0",
 		"eslint-config-prettier": "^9.1.0",
-		"eslint-plugin-svelte": "^2.35.1",
+		"eslint-plugin-svelte": "^2.36.0-next.4",
 		"postcss": "^8.4.32",
 		"postcss-load-config": "^5.0.2",
 		"prettier": "^3.1.1",
 		"prettier-plugin-svelte": "^3.1.2",
 		"prettier-plugin-tailwindcss": "^0.5.9",
-		"svelte": "^4.2.7",
+		"svelte": "^5.0.0-next.1",
 		"svelte-check": "^3.6.0",
+		"svelte-radix": "^1.0.1",
 		"tailwindcss": "^3.3.6",
 		"tslib": "^2.4.1",
 		"typescript": "^5.0.0",
@@ -36,9 +37,11 @@
 	},
 	"type": "module",
 	"dependencies": {
+		"@internationalized/date": "^3.5.2",
 		"bits-ui": "^0.18.2",
 		"clsx": "^2.1.0",
-		"lucide-svelte": "^0.338.0",
+		"lucide-svelte": "^0.341.0",
+		"mode-watcher": "^0.2.1",
 		"tailwind-merge": "^2.2.1",
 		"tailwind-variants": "^0.2.0"
 	}
