@@ -1,5 +1,7 @@
-<script>
+<script lang="ts">
   import '../app.pcss';
+  import { ModeWatcher } from "mode-watcher";
 </script>
 
+<ModeWatcher />
 <slot />
